mood-music-app-main/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
itsdangerous==2.1.2
slowapi==0.1.8
python-dotenv==1.0.0
httpx==0.25.2
pydantic==2.5.2
python-multipart==0.0.6
starlette==0.27.0
click==8.1.8
opencv-python==4.8.1.78
numpy==1.26.2
requests==2.31.0
deepface==0.0.79
google-generativeai==0.3.2
